fix(class-validator): require ApiGatewayEvent
This solves a type problem related to compose: TypeScript cannot infer generics of composited functions correctly, so we need to get rid of generics to make composition easier.

BREAKING CHANGE: Before, class-validator middleware could be used for handler of all kinds of events. Now it can only be used for APIGatewayEvents. This should not cause much trouble, as the middleware already depended on the body being in the same form as in an APIGatewayEvent.