feat(json-deserilalizer): initial implementation of json- deserializer package
This implements a json-deserializer middleware, it expects an object extending ApiGatewayProxyEvent and returns an ApiGatewayProxyObjectEvent which replaces the string body for an object of type Record<string, unknown>.

It introduces the concept of a RequestBodyNotJsonError to be used to return an appropriate 400 error back if passed a non json payload is passed.

If no body or a non-json header is passed it will pass null as the body property on the event passed to the underlying handler.